+++ changelog.md
@@ -1,4 +1,7 @@
 # Changelog
 
-* Deleted './src/output/result-control.js'
-	* It has been split into different files.
+**./src/common/sub-interface/**
+* When displaying help on the command line, all command and option descriptions are now lowercase.
+* This affects the files:
+	* cmd-desc.js
+	* option-desc.js